# go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-Performance Go Control Plane \u0026 Observability API for Algorithmic Trading.",
        "title": "Trading Observability API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/api/alpaca/account": {
            "get": {
                "description": "Retrieve broker account info (balance, buying power, equity)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alpaca"
                ],
                "summary": "Get Alpaca Account Info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key for authorization",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/alpaca/bars": {
            "get": {
                "description": "Query historical market candlestick bars from Alpaca",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alpaca"
                ],
                "summary": "Get Market Bars",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key for authorization",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Stock symbol",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Number of bars to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"1Min\"",
                        "description": "Bar size (1Min/5Min/1Hour/1Day)",
                        "name": "time_frame",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/alpaca/orders": {
            "get": {
                "description": "List recent orders with current fulfillment status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alpaca"
                ],
                "summary": "List Alpaca Orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key for authorization",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Cancel all pending/active orders",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alpaca"
                ],
                "summary": "Cancel All Alpaca Orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key for authorization",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/alpaca/orders/limit": {
            "post": {
                "description": "Submit a new limit order to buy/sell assets at a specified price",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alpaca"
                ],
                "summary": "Place Alpaca Limit Order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key for authorization",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Limit order parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.AlpacaLimitOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/alpaca/orders/market": {
            "post": {
                "description": "Submit a new market order to buy/sell assets immediately",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alpaca"
                ],
                "summary": "Place Alpaca Market Order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key for authorization",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Market order parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.AlpacaMarketOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/alpaca/orders/{order_id}": {
            "delete": {
                "description": "Cancel a specific pending order by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alpaca"
                ],
                "summary": "Cancel Alpaca Order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key for authorization",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Order UUID",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/alpaca/positions": {
            "get": {
                "description": "List all open investment positions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alpaca"
                ],
                "summary": "Get Alpaca Positions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key for authorization",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Liquidate all open positions (market order sells for all holds)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alpaca"
                ],
                "summary": "Close All Positions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key for authorization",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/system/alerts/acknowledge/{alert_id}": {
            "post": {
                "description": "Acknowledge a specific alert notification by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Acknowledge Alert",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key for authorization",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Alert identifier",
                        "name": "alert_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/system/components": {
            "get": {
                "description": "Check detailed status of subsystem components",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get System Components",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key for authorization",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/system/incidents/{incident_id}/acknowledge": {
            "post": {
                "description": "Mark an incident as acknowledged and assign an owner",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Acknowledge Incident",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key for authorization",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Incident UUID",
                        "name": "incident_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of the owner resolving the incident",
                        "name": "owner",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "http.AlpacaLimitOrderRequest": {
            "type": "object",
            "properties": {
                "limit_price": {
                    "type": "number",
                    "example": 150.25
                },
                "qty": {
                    "type": "number",
                    "example": 10
                },
                "side": {
                    "type": "string",
                    "example": "buy"
                },
                "symbol": {
                    "type": "string",
                    "example": "AAPL"
                },
                "time_in_force": {
                    "type": "string",
                    "example": "day"
                }
            }
        },
        "http.AlpacaMarketOrderRequest": {
            "type": "object",
            "properties": {
                "qty": {
                    "type": "number",
                    "example": 10
                },
                "side": {
                    "type": "string",
                    "example": "buy"
                },
                "symbol": {
                    "type": "string",
                    "example": "AAPL"
                },
                "time_in_force": {
                    "type": "string",
                    "example": "day"
                }
            }
        },
...
    }
}